Back/requirements.txt
asgiref==3.7.2
Django==4.2.2
django-cors-headers==4.0.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
Pillow==9.5.0
PyJWT==1.7.1
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
